monitoring/pyproject.toml
[tool.poetry]
name = "monitoring"
version = "0.1.0"
description = ""
authors = ["xrsrke <xariusdrake@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
evidently=">=0.3.2.1"
fastapi=">=0.95.2"
mongo=">=0.2.0"
pandas=">=2.0.2"
pydantic=">=1.10.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
